refactor: update import paths and implement dynamic caching patterns in service worker for improved asset management
