+++ src/app.js
@@ -1,4 +1,4 @@
-import { INCLUDE_TRAVELLERS_KEY } from '../constants.js';
+import { INCLUDE_TRAVELLERS_KEY } from './constants.js';
 import { renderSetupInfo, setupGrimoire, updateGrimoire } from './grimoire.js';
 import { repositionPlayers } from './ui/layout.js';
 import { processScriptData } from './script.js';
